+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 12.91.0
+
+* `[terminal/tty]` Added method `IsTMUX`
+* `[terminal/tty]` Added more usage examples
+* `[timeutil]` `ParseDuration` now returns `time.Duration` instead of seconds
+
 ### 12.90.1
 
 * `[pager]` Use `more -f` by default